refactor: moved pack directory into src
